Create helper module to help test dependency resolution tests
